fix: note index doesn't move with viewport
