+++ packages/blocks/src/page-block/edgeless/edgeless-page-block.ts
@@ -80,7 +80,7 @@ import { RangeSynchronizer } from '../text-selection/range-synchronizer.js';
 import { NoteCut } from './components/note-cut/index.js';
 import { EdgelessNotesStatus } from './components/notes-status.js';
 import { EdgelessToolbar } from './components/toolbar/edgeless-toolbar.js';
-import { readImageSize } from './components/utils.js';
+import { readImageSize, updateNotesPoision } from './components/utils.js';
 import { ZoomBarToggleButton } from './components/zoom/zoom-bar-toggle-button.js';
 import {
   EdgelessZoomToolbar,
@@ -538,6 +538,8 @@ export class EdgelessPageBlockComponent
         if (this.selection.selectedBlocks.length) {
           this.selection.setSelectedBlocks([...this.selection.selectedBlocks]);
         }
+
+        updateNotesPoision(this, this.notes);
       })
     );
     _disposables.add(
